Compute the relative difference

Computes the relative difference statistic of the form (est - pop)/ pop, which is equivalent to the form est/pop - 1. If matrices are supplied then an equivalent matrix variant will be used of the form (est - pop) * solve(pop). Values closer to 0 indicate better relative parameter recovery. Note that for single variable inputs this is equivalent to bias(..., type = 'relative').

RD(est, pop, as.vector = TRUE, unname = FALSE)

Arguments

est

a numeric vector, matrix/data.frame, or list containing the parameter estimates

pop

a numeric vector or matrix containing the true parameter values. Must be of comparable dimension to est

as.vector

logical; always wrap the result in a as.vector function before returning?

unname

logical; apply unname to the results to remove any variable names?

Value

returns a vector or matrix depending on the inputs and whether as.vector was used

References

Chalmers, R. P., & Adkins, M. C. (2020). Writing Effective and Reliable Monte Carlo Simulations with the SimDesign Package. The Quantitative Methods for Psychology, 16(4), 248-280. doi:10.20982/tqmp.16.4.p248

Sigal, M. J., & Chalmers, R. P. (2016). Play it again: Teaching statistics with Monte Carlo simulation. Journal of Statistics Education, 24(3), 136-156. doi:10.1080/10691898.2016.1246953

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

# vector
pop <- seq(1, 100, length.out=9)
est1 <- pop + rnorm(9, 0, .2)
(rds <- RD(est1, pop))
#> [1] -0.1221416359 -0.0001433052  0.0014094503  0.0001416364  0.0046819446
#> [6]  0.0029448524  0.0022352420 -0.0015123786  0.0017933783
summary(rds)
#>       Min.    1st Qu.     Median       Mean    3rd Qu.       Max. 
#> -0.1221416 -0.0001433  0.0014095 -0.0122879  0.0022352  0.0046819 

# matrix
pop <- matrix(c(1:8, 10), 3, 3)
est2 <- pop + rnorm(9, 0, .2)
RD(est2, pop, as.vector = FALSE)
#>            [,1]       [,2]      [,3]
#> [1,] 0.04717509 -0.2042294 0.1239549
#> [2,] 0.17461297 -0.6481818 0.4303047
#> [3,] 0.38129425 -1.0145981 0.5729735
(rds <- RD(est2, pop))
#> [1]  0.04717509  0.17461297  0.38129425 -0.20422939 -0.64818184 -1.01459805
#> [7]  0.12395488  0.43030474  0.57297354
summary(rds)
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> -1.01460 -0.20423  0.12395 -0.01519  0.38129  0.57297