Form Column Standard Deviation and Variances

Form column standard deviation and variances for numeric arrays (or data frames).

colVars(x, na.rm = FALSE, unname = FALSE)

colSDs(x, na.rm = FALSE, unname = FALSE)

Arguments

x

an array of two dimensions containing numeric, complex, integer or logical values, or a numeric data frame

na.rm

logical; remove missing values in each respective column?

unname

logical; apply unname to the results to remove any variable names?

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

results <- matrix(rnorm(100), ncol=4)
colnames(results) <- paste0('stat', 1:4)

colVars(results)
#>     stat1     stat2     stat3     stat4 
#> 0.7549535 0.4807686 0.7438587 1.4105222 
colSDs(results)
#>     stat1     stat2     stat3     stat4 
#> 0.8688806 0.6933748 0.8624725 1.1876541 

results[1,1] <- NA
colSDs(results)
#>     stat1     stat2     stat3     stat4 
#>        NA 0.6933748 0.8624725 1.1876541 
colSDs(results, na.rm=TRUE)
#>     stat1     stat2     stat3     stat4 
#> 0.8857748 0.6933748 0.8624725 1.1876541 
colSDs(results, na.rm=TRUE, unname=TRUE)
#> [1] 0.8857748 0.6933748 0.8624725 1.1876541