Compute univariate descriptive statistics

Function returns univariate data summaries for each variable supplied, however discrete and continuous variables are treated separately. Conditional statistics are evaluated internally using the by function, however group-level specifications are declared using the dplyr style inputs; specifically, group_by. Quantitative/continuous variable information is kept distinct in the output, while discrete observations can be returned using the explicit discrete argument.

descript(
  df,
  funs = c(n = length, mean = mean, trimmed = function(x) mean(x, trim = 0.1), sd = sd,
    mad = mad, skewness = EnvStats::skewness, kurtosis = EnvStats::kurtosis, min = min,
    Q_25 = function(x) quantile(x, 0.25), Q_50 = median, Q_75 = function(x) quantile(x,
    0.75), max = max),
  discrete = FALSE
)

Arguments

df

a data.frame or tibble containing the variables of interest. Note that factor and character vectors will be treated as discrete observations

funs

functions to apply when discrete = FALSE. Can be modified by the user to include or exclude further functions, however each supplied function must return a scalar.

Default functions returns:

n

number of non-missing observations

mean

mean

trimmed

trimmed mean (10%)

sd

standard deviation

mad

mean absolute deviation

skewness

skewness (from EnvStats)

kurtosis

kurtosis (from EnvStats)

min

minimum

Q_25

25% quantile

Q_50

50% quantile (a.k.a., the median)

Q_75

75% quantile

max

maximum

discrete

logical; include summary statistics for discrete variables only? If TRUE then only count and proportion information will be returned

See also

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

data(mtcars)
fmtcars <- within(mtcars, {
  cyl <- factor(cyl)
  am <- factor(am, labels=c('automatic', 'manual'))
  vs <- factor(vs)
})

# with and without factor variables
mtcars |> descript()  # omitted
#> Error in loadNamespace(x): there is no package called ‘EnvStats’
fmtcars |> descript()  # omitted
#> Error in loadNamespace(x): there is no package called ‘EnvStats’

# discrete variables only
fmtcars |> descript(discrete=TRUE)
#> $cyl
#> # A tibble: 3 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 4         11      0.344
#> 2 6          7      0.219
#> 3 8         14      0.438
#> 
#> $vs
#> # A tibble: 2 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 0         18      0.562
#> 2 1         14      0.438
#> 
#> $am
#> # A tibble: 2 × 3
#>   values    count proportion
#>   <fct>     <int>      <dbl>
#> 1 automatic    19      0.594
#> 2 manual       13      0.406
#> 

# conditioning
fmtcars |> group_by(cyl) |> descript()
#> Error in loadNamespace(x): there is no package called ‘EnvStats’
fmtcars |> group_by(cyl, am) |> descript()
#> Error in loadNamespace(x): there is no package called ‘EnvStats’

# conditioning also works with group_by()
fmtcars |> group_by(cyl) |> descript(discrete=TRUE)
#> cyl: 4
#> $vs
#> # A tibble: 2 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 0          1     0.0909
#> 2 1         10     0.909 
#> 
#> $am
#> # A tibble: 2 × 3
#>   values    count proportion
#>   <fct>     <int>      <dbl>
#> 1 automatic     3      0.273
#> 2 manual        8      0.727
#> 
#> ------------------------------------------------------------ 
#> cyl: 6
#> $vs
#> # A tibble: 2 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 0          3      0.429
#> 2 1          4      0.571
#> 
#> $am
#> # A tibble: 2 × 3
#>   values    count proportion
#>   <fct>     <int>      <dbl>
#> 1 automatic     4      0.571
#> 2 manual        3      0.429
#> 
#> ------------------------------------------------------------ 
#> cyl: 8
#> $vs
#> # A tibble: 2 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 0         14          1
#> 2 1          0          0
#> 
#> $am
#> # A tibble: 2 × 3
#>   values    count proportion
#>   <fct>     <int>      <dbl>
#> 1 automatic    12      0.857
#> 2 manual        2      0.143
#> 
fmtcars |> group_by(cyl, am) |> descript(discrete=TRUE)
#> cyl: 4
#> am: automatic
#> $vs
#> # A tibble: 2 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 0          0          0
#> 2 1          3          1
#> 
#> ------------------------------------------------------------ 
#> cyl: 6
#> am: automatic
#> $vs
#> # A tibble: 2 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 0          0          0
#> 2 1          4          1
#> 
#> ------------------------------------------------------------ 
#> cyl: 8
#> am: automatic
#> $vs
#> # A tibble: 2 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 0         12          1
#> 2 1          0          0
#> 
#> ------------------------------------------------------------ 
#> cyl: 4
#> am: manual
#> $vs
#> # A tibble: 2 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 0          1      0.125
#> 2 1          7      0.875
#> 
#> ------------------------------------------------------------ 
#> cyl: 6
#> am: manual
#> $vs
#> # A tibble: 2 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 0          3          1
#> 2 1          0          0
#> 
#> ------------------------------------------------------------ 
#> cyl: 8
#> am: manual
#> $vs
#> # A tibble: 2 × 3
#>   values count proportion
#>   <fct>  <int>      <dbl>
#> 1 0          2          1
#> 2 1          0          0
#>